Rewrite of tests.
